+++ src/Components/CreateNewUser/CreateNewUser.css
@@ -15,7 +15,7 @@ body{
     color: white;
 }
 
-.btn-iniciar-sesion{
+.btn-crear-cuenta{
    background-color: var(--main-white);
 }
