+++ src/Components/Pages/CreateNewUserPage.jsx
@@ -1,0 +1,10 @@
+import React from 'react'
+import CreateNewUser from '../CreateNewUser/CreateNewUser';
+
+const CreateNewUserPage = () => {
+  return (
+    <CreateNewUser />
+  )
+}
+
+export default CreateNewUserPage
