+++ src/Components/Pages/LoginPage.jsx
@@ -1,5 +1,6 @@
 import React from 'react'
-import Login from '../Login/Login';
+import Login from '../Login/Login'
+
 const LoginPage = () => {
   return (
     <Login />
