+++ src/index.js
@@ -3,7 +3,8 @@ import ReactDOM from 'react-dom/client';
 // import components
 import App from './App';
 // import pages
-import LoginPage from './Components/Pages/LoginPage'
+import CreateNewUserPage from './Components/Pages/CreateNewUserPage'
+import LoginPage from './Components/Pages/LoginPage';
 import {
   BrowserRouter,
   Routes,
@@ -18,7 +19,8 @@ root.render(
     <BrowserRouter>
     <Routes>
       <Route path='/' element={<App />} />
-      <Route path='/Login' element={< LoginPage />} />
+      <Route path='/CreateNewUser' element={ <CreateNewUserPage /> } />
+      <Route path='/LogIn' element={ <LoginPage /> } />
     </Routes>
     </BrowserRouter>
   </React.StrictMode>
